# console/console_nakama.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Nakama console API v2",
    "version": "2.0",
    "contact": {
      "name": "The Nakama Authors \u0026 Contributors",
      "url": "https://github.com/heroiclabs/nakama",
      "email": "hello@heroiclabs.com"
    }
  },
  "host": "127.0.0.1:7351",
  "schemes": [
    "http"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v2/console/runtime": {
      "get": {
        "summary": "Get runtime info",
        "operationId": "GetRuntime",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/RuntimeInfo"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpc.Status"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Any": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "ModuleInfo": {
      "type": "object",
      "properties": {
        "path": {
          "type": "string",
          "title": "Module path"
        },
        "mod_time": {
          "type": "string",
          "format": "date-time",
          "title": "Module last modified date"
        }
      },
      "title": "Module information"
    },
...
    "RuntimeInfo": {
      "type": "object",
      "properties": {
        "go_rpc_functions": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "Lua registered RPC functions\nrepeated string lua_rpc_functions = 1;\nGo registered RPC functions"
        },
        "js_rpc_functions": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "title": "JavaScript registered RPC functions"
        },
        "go_modules": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ModuleInfo"
          },
          "title": "Go loaded modules"
        },
        "js_modules": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/ModuleInfo"
          },
          "title": "Lua loaded modules\nrepeated ModuleInfo lua_modules = 5;\nJavaScript loaded modules"
        }
      },
      "title": "Runtime information"
    },
...
    "rpc.Status": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/Any"
          }
        }
      }
    }
  },
  "securityDefinitions": {
    "BasicAuth": {
      "type": "basic"
    },
    "BearerJwt": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "BearerJwt": []
    }
  ],
...
}